Restructure modules organization.
