+++ erp/imports/mapper/vaccination.py
@@ -4,8 +4,8 @@ from django.contrib.gis.geos import Point
 from django.db.utils import DataError
 
 from core.lib import text
-from erp.import_datasets.base_mapper import BaseRecordMapper
-from erp.import_datasets.fetcher_strategy import Fetcher
+from erp.imports.mapper.base import BaseRecordMapper
+from erp.imports.fetcher import Fetcher
 from erp.models import Accessibilite, Commune, Erp
 from erp.provider import arrondissements
